refactor(router): clarify RoutingHost boundaries
