+++ lib/llm/src/kv_router/routing_host/kv_selection.rs
@@ -16,7 +16,7 @@ use dynamo_runtime::{dynamo_nvtx_range, pipeline::Error};
 use crate::{
     kv_router::{
         FindBestMatchAdmission, FindBestMatchInnerOutcome, FindBestMatchOutcome,
-        push_router::RoutingHost,
+        routing_host::RoutingHost,
     },
     local_model::runtime_config::ModelRuntimeConfig,
     preprocessor::PreprocessedRequest,
