+++ lib/llm/src/kv_router/routing_host/occupancy.rs
